+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.1] - 2025-02-25
+
+### Changed
+
+- Mod logo
+
+### Fixed
+
+- Advanced Random Trigger popup crash
+- Sequence Trigger popup crash
+- Collision Trigger popup crash (#5)
+- Named IDs popup items weird positioning bug ($4)
+
 ## [1.0.0] - 2025-02-22
 
 Initial Commit